Implementing Edge-induced subgraph isomorphism fix
Initially using GraphMatcher.subgraph_is_isomorphic() was causing a bug. Inspecting the documentation this bug came from the fact that the VF2 algorithm can only find subgraphs with respect to vertex-induced subgraphs. Translating the Host and PotentialSubgraph to the nx.line_graph() of the respective graphs fixes this and allows the VF2 algorithm to find edge-induced subgraph isomorphisms.