Added largest subgraph of intersection
Implemented the search for the largest subgraph of the red and blue subgraphs of each coloring. the algorithm then searches for the list of unique subgraphs of this type.
This version also adds separate functions draw_graph_list and save_graph_list that do what they sound like they do. These are parameterized and results in a square image that’s as small as feasible for the given list.
Currently, I’m ignoring the K_1,1 bug that stems from there only being one red-blue coloring up to isomorphism of red and blue. The bug is specifically that the total poset is empty, and the largest intersection is empty. There very well may also be a bug that 2P_2 is not a subgraph of P_3 that I notice from the total poset structure of K_2,2