Bugfix for colorings and subgraphs
This version also includes cells that inspect the first few: paths, cycles, complete graphs, and complete bipartite graphs.

This version now iterates through all possible number of combinations of edges to choose those to be red. This also includes choosing no edges to be colored red. This version now also uses the nx.algorithms.isomorphism.GraphMatcher(G,H).subgraph_is_monomorphic() because the definition of monomorphic in networkx is the definition of subgraph isomorphic in the graph theory literature.