Implemented “union”
In this version, I add the implementation for determining the smallest subgraph that contains both a copy of the red subgraph, and the blue subgraph corresponding to a particular red-blue coloring of a graph. This does not, however, lead closer to the Down Arrow set, however. This will be coming in a future update