Union Intersection Fix
This implements the union and intersection fix. This will inherently use more memory so as to not have to find the set of all unique subgraphs of both the red and blue coloring multiple times when determining the intersection. Still need to implement a way to save the union of the colorings, and to intersect the result with an ongoing intersection.