Merge
This is a complete rewrite into one file that generates the unique red subgraphs, unique subgraphs, poset of subgraphs, determines the down arrow set, and ideals that generate the down arrow set of a given graph. Each step in the process (minus determining the ideals and making the poset) are parallelized. Determining the poset of subgraphs still needs to be parallelized, and not paralized.