Comments
This adds comments throughout, as well as tentative change to isomorphism checks in intersection. Since this method will only ever be called after the poset is given, the structure of the graph is deterministic and it should be quicker to check if the graph6 string is equal instead of riggorously determining an isomorphism.